Fix miseq multiqc file name split
If miseq the flowcell id contains a "-". The multiqc file name is split on "-" to capture the project id but the presense of this extra dash means this is now incorrect. 

Fix takes the html name, reverses it, cuts at "-" and takes the second and third elements after cutting, then reverses these again to give the correct file name (this makes sure that it is always the "-multiqc.html" element being removed regardless of the number of "-" present). Tested on NextSeq and MiSeq project names.